Move ToggleLightAction to Middleware. Add UpdatedLightStateAction.
