+++ lib/redux/middleware.dart
@@ -77,8 +77,37 @@ void lightsApiMiddleware(Store<AppState> store, dynamic action, NextDispatcher n
     }
   }
 
+  void toggleLight(String lightId) async {
+    final Map<String, bool> lightsStateCache = <String, bool>{};
+    final ip = store.state.config.ipAddress;
+    final applicationKey = store.state.config.username;
+    var state = lightsStateCache[lightId];
+
+    final Uri uri = Uri.parse('https://$ip/clip/v2/resource/light/$lightId');
+    var body = json.encode({
+      "on": {"on" : state}
+    });
+
+    final Map<String, String> headers = { 'hue-application-key' : applicationKey };
+
+    try {
+      await http.put(uri, body: body, headers: headers);
+      if (state != null) {
+        lightsStateCache[lightId] = state;
+        store.dispatch(UpdatedLightStateAction(lightsStateCache));
+      }
+    } catch (error) {
+      print(error);
+    }
+  }
+
   if (action is GetLightsStateAction) {
     getLightsState();
+  }
+
+  if (action is ToggleLightAction) {
+    ToggleLightAction a = action;
+    toggleLight(a.lightId);
   }
 
   next(action);
