+++ lib/redux/reducers.dart
@@ -40,5 +40,14 @@ AppState appStateReducer(AppState state, action) {
     return AppState(ApplicationState.fetchedLightsState, state.config, state.deviceList, a.lightsStateCache);
   }
 
+  if (action is ToggleLightAction) {
+    return AppState(ApplicationState.userInteractive, state.config, state.deviceList, state.lightsStateCache);
+  }
+
+  if (action is UpdatedLightStateAction) {
+    final UpdatedLightStateAction a = action;
+    return AppState(ApplicationState.userInteractive, state.config, state.deviceList, a.lightsStateCache);
+  }
+
   return state;
 }
